Initialize block mgr limit based on the query limit, not current consumption.
This was always the intent and the code is wrong. The process mem tracker
is what is responsible for coordinating across queries.

We see this cause failures in the perf runs when we start up the next query
right after the current one finishes. This happens before the current one
has cleaned up, causing the new query to run with very little memory.

Change-Id: I1e25e55959c14c58789eaa38a122470e14761767
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4769
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: jenkins