Temp fix for IMPALA-1488: disable spilling hash tables with matches in right joins.
Right joins (right outer, right semi, right anti, and full outer) depend on the matched
flag for the build side, information stored in the hash tables. Thus, in right joins
if we spill a hash table that has matches, then we are going to lose this information
and return wrong results.

This patch adds a flag in the hash table which is set in case of right joins that had
at least one match. Then, the SpillPartition() algorithm won't spill partitions that
had hash table matches. If there are no partitions to spill the query will gracefully
fail with OOM.

Change-Id: I736400768529019bb10c2541de552d958eb90044
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5306
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: jenkins
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5335