IMPALA-1507: Wrong DCHECK when trying to spill just initialized PAGG partition
When a user would set very low mem limit and run even simple aggregation query,
there were cases where there was not enough memory to allocate even the initial
set of partitions for PAGG. In that cases we were trying to spill partitions.
But if that was the initial set of partitions created, then there was a DCHECK
that was checking if the streams were using small_buffers (the default), without
checking if there were actually any rows in it. There result was a crash without
any meaningful error message.

This patch fixes the DCHECK, and queries gracefully fail with OOM.

Change-Id: I0446895330e2ffb928becf8f09e93435a500ab9e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5521
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: jenkins