IMPALA-1590: DCHECK(pinned) in Sorter when running concurrent queries
Sorter::PrepareRead() assumed that all the var-len blocks would Pin() successfully,
but that is not the case when there are multiple concurrent queries. In particular, if
the number of var-len blocks is larger than the minimum blocks reserved (6 in the case
of var len slots), then Pin() may fail, probably because of OOM due to concurrent queries.

This patch keeps the DCHECK when the number of var-len blocks is smaller than the reserved
blocks, but fails the query with OOM otherwise.

Change-Id: I342e121afaf88e72b2ff645991685a64b0d2dde5
Reviewed-on: http://gerrit.cloudera.org:8080/159
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins