IMPALA-1623: unix_timestamp() does not return correct time
The compiler was choosing TimestampValue(bool) as the constructor for a
call of TimestampValue(*timestamp_value). It turns out that the
unix_timestamp() function didn't need to construct a TimestampValue
anyways but I added the pointer constructor anyways so this won't happen
in the future.

Change-Id: I311f2c9f10ae0aa8c2ce3f5babaeacb258ffdb0f
(cherry picked from commit 436bdb9f068803b422795846c35d9189c83e1da3)