IMPALA-1663: Disable background catalog loads by default
Trying to load the entire catalog in the background on startup leads to
excessive network usage for large catalogs or clusters. We should
default --load_catalog_in_background=false for the catalog server and
pay the hit on the first table access.

Ideally we would be smarter here: do better rate-limiting of the catalog
load, and stop background loading if the catalog gets too big (see
IMPALA-804). But for now, disabling the background load is the easiest
thing to do.

Change-Id: I81a6d0321491ec726fc03fae271d0e59f694adbf
Reviewed-on: http://gerrit.cloudera.org:8080/126
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins