IMPALA-1668: Fix leak of transport objects in TSaslServerTransport::Factory
Thrift servers use both an input and an output transport to communicate with
clients. In principal, these can be different, but for SASL clients we require them
to be the same so that the authentication state is identical for communication in
both directions. In order to do this, we cache the transport that we return in a map
keyed by the transport argument to this method. Then if there are two successive
calls to getTransport() with the same transport, we are sure to return the same
wrapped transport both times.

However, the cache map would retain references to all the transports it ever
created. Instead, we remove an entry in the map after it has been found for the first
time, that is, after the second call to getTransport() with the same argument. That
matches the calling pattern in TThreadedServer and TThreadPoolServer, which both call
getTransport() twice in succession when a connection is established, and then never
again. This is obviously brittle (what if for some reason getTransport() is called a
third time?) but for our usage of Thrift it's a tolerable band-aid.

An alternative approach is to use the 'custom deleter' feature of shared_ptr to
ensure that when ret_transport is eventually deleted, its corresponding map entry is
removed. That is likely to be error prone given the locking involved; for now we go
with the simple solution.

Change-Id: I00f089a66f9993c7f4cbbe97e241b6e101488a2e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5781
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: jenkins
(cherry picked from commit bc14b72450a2093d40fff156918048d3865c379c)