Disable ERROR double-logging to stdout
Glog has a useful feature where ERROR messages are always at least
written to stdout, no matter what the rest of the logging setup
is. However, we don't rely on stdout in our deployments, and for local
tests this means that Java LOG.error() messages are always written to
stdout. This affects some of our tests which deliberately hit error
paths, and produce a lot of output (Sentry is the prime example). So
let's always disable double-logging (it was mostly disabled before).

Change-Id: Ife6f8e37b1b8eed0fee887c296c40f8ddb3a73e7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5787
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: jenkins