IMPALA-1622: fix overflow in StringParser::StringToFloatInternal()
This bug was caused by overflowing the 'remainder' int when parsing a
floating point number with many digits after the decimal. To fix this,
this patch changes the function to stop computing remainder after
seeing 18 significant figures (i.e., digits not including leading 0s),
which prevents it from overflowing. This is safe to do according to
the IEEE binary64 (i.e. double) spec.

Change-Id: I98ec68bcae96a6b647bc274e28d3f6c2f6067769
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5879
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins