IMPALA-1691: Remove HMS partition from HdfsPartition
For convenience, every HdfsPartition object would cache its HMS
counterpart, which was needed for alteration RPCs to the HMS. However,
this leaves us vulnerable to excessive memory consumption in the Hive
Thrift representation of a partition (since Hive itself doesn't cache
these structures in memory, they care less about their size), and we
found that each partition object stored the full list of
FieldSchemas. This is redundant information since the FieldSchema is
consistent across all partitions.

Instead, this patch changes HdfsPartition to store any unrecoverable
information from the HMS partition object, and then use that to
construct an HMS partition as needed. This changes the order of
operations when a partition is altered, which used to be (roughly):

HMSPartition p = hdfsPart.getHmsPartition();
// alter p by setting location or file format etc.
alterPartitionInHms(p);
hdfsPart.markDirty();

and then relying on the dirty flag / update mechanism to refresh the
hdfsPart object.

Now the operations are:

// alter hdfsPart *in place*
alterPartitionInHms(p.getHmsPartition());
hdfsPart.markDirty();

So hdfsPart is altered before the HMS is updated.

Change-Id: Iae60b3c2a226ca03b94f8cea31c3c9aa4dd907a9
Reviewed-on: http://gerrit.cloudera.org:8080/36
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins