IMPALA-1209: LOAD DATA support between HDFS encryption zones
LOAD DATA will fail if the target data directory and the table
directory are on different HDFS encryption zones. Currently we
move the target data into the table directory, but this metadata
(NN only) operation will fail with a permissions error. The data
may need to be decrypted (if source data is encrypted) and/or
encrypted (if target table directory is encrypted). This is
accomplished by copying the files rather than moving them.

Change-Id: I07578702dc2e85d183b52272393092966270e159
Reviewed-on: http://gerrit.cloudera.org:8080/81
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins