IMPALA-1655: Probe filters were being attached even if partitions were spilled
PHJ can hit OOM while building a hash table. In that case it will stop building the HT
and try to spill a partition. Unfortunately, if PHJ had determined that it should build
probe filters (ie. the input is small enough), it would attach those incomplete probe
filters ignoring the fact that some tuples were not seen. Those incomplete probe filters
would be used by the scanners producing wrong results as more tuples would be filtered
out.

This patch solves the problem by setting that probe filters should not be used when a
hash table fails to build. It also adds memory scaling tests for TPC-H Q9 and Q21 that
expose this problem.

Change-Id: I58d97aa0fd3dce92f67482c301178877177df6dd
Reviewed-on: http://gerrit.cloudera.org:8080/99
Reviewed-by: Ippokratis Pandis <ippokratis@gmail.com>
Tested-by: Internal Jenkins