Fix node output cardinality for broadcast recipients
Broadcast exchanges send the same data to all recipients, rather than
partitioning the output. That means that the output cardinality reported
by each exchange node on the recipient plan fragment will be the same,
and so summing across all instances to get the output cardinality gives
an artificially high number. This patch fixes the cardinality value for
broadcasts, however we should note that this now *understates* the total
rows sent across the wire. Since the plan view is kind of a flattened
logical view, this is a clearer way to show the logical number of tuples
that flow up the plan.

Change-Id: I7ccd993e922b2d1dfa292ed30fc7fc1a958af882
Reviewed-on: http://gerrit.cloudera.org:8080/103
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins