Enable local filesystem tables
The S3 work really enabled any Hadoop FileSystem to work with Impala,
but a small tweak is needed for LocalFileSystem due to how the Hadoop
Path code deals with URIs that don't have an authority component.

While we aren't claiming support for arbitrary FileSystem's at this
tiem, it is useful to test this.  Since the S3 testing is done as a
nightly test rather than pre-checkin, we can use the LocalFileSystem to
regression test that:

1) Impala can access table data living on a secondary filesystem,
   i.e. not the filesystem specified by fs.defaultFS.
2) Impala does not make assumptions that the filesystem has type
   DistributedFileSystem.

Change-Id: Ie9b858ea440c9b3b332602e034c8052b168c57da
Reviewed-on: http://gerrit.cloudera.org:8080/121
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: Internal Jenkins